fix(presenter): streamline video frame handling in ScreenCaptureService and ScreenShareService
